+++ src/compass/utils/geometry_utils.py
@@ -7,6 +7,7 @@
 
 
 import numpy as np
+import isce3
 
 
 def los2orbit_azimuth_angle(los_az_angle, look_direction='right'):
@@ -265,3 +266,124 @@ def get_unit_vector4component_of_interest(los_inc_angle, los_az_angle, comp='enu
         ]
 
     return unit_vec
+
+def enu2rgaz(radargrid_ref, orbit, ellipsoid,
+             lon_arr, lat_arr, hgt_arr,
+             e_arr, n_arr, u_arr):
+    '''
+    Convert ENU displacement into range / azimuth displacement
+
+    Parameters
+    ----------
+    radargrid_ref: isce3.product.RadarGridParameters
+        Radargrid of the burst
+    orbit: isce3.core.Orbit
+        Orbit of the burst
+    ellipsoid: isce3.core.Ellipsoid
+        Ellipsoid definition
+
+    lon_arr, lat_arr, hgt_arr: np.nadrray
+        Arrays for lonfigute, latitude, and height.
+        Units for longitude and latitude are degree; unit for height is meters.
+    
+    e_arr, n_arr, u_arr: np.ndarray
+        Displacement in east, north, and up direction in meters
+
+    Returns
+    -------
+    rg_arr: np.ndarray
+        Displacement in slant range direction in meters.
+    az_arr: np.ndarray
+        Displacement in azimuth direction in seconds.
+    '''
+    shape_arr = lon_arr.shape
+    rg_arr = np.zeros(shape_arr)
+    az_arr = np.zeros(shape_arr)
+
+    # Calculate the ENU vector in ECEF
+    for i, lon_deg in enumerate(np.nditer(lon_arr)):
+        index_arr = np.unravel_index(i, lon_arr.shape)
+        lat_deg = lat_arr[index_arr]
+        hgt = hgt_arr[index_arr]
+
+        vec_e, vec_n, vec_u = get_enu_vector_ecef(lon_deg, lat_deg)
+
+        llh_ref = np.array([np.deg2rad(lon_deg),
+                               np.deg2rad(lat_deg),
+                               hgt])
+
+        xyz_before = ellipsoid.lon_lat_to_xyz(llh_ref)
+        xyz_set = (xyz_before + vec_e * e_arr[index_arr]
+                              + vec_n * n_arr[index_arr]
+                              + vec_u * u_arr[index_arr])
+        llh_displaced = ellipsoid.xyz_to_lon_lat(xyz_set)
+
+        aztime_ref, slant_range_ref =\
+            isce3.geometry.geo2rdr(llh_ref,
+                                   ellipsoid,
+                                   orbit,
+                                   isce3.core.LUT2d(),
+                                   radargrid_ref.wavelength,
+                                   radargrid_ref.lookside,
+                                   threshold=1.0e-10, maxiter=50, delta_range=10.0)
+
+        aztime_displaced, slant_range_displaced =\
+            isce3.geometry.geo2rdr(llh_displaced,
+                                   ellipsoid,
+                                   orbit,
+                                   isce3.core.LUT2d(),
+                                   radargrid_ref.wavelength,
+                                   radargrid_ref.lookside,
+                                   threshold=1.0e-10, maxiter=50, delta_range=10.0)
+
+        rg_arr[index_arr] = slant_range_displaced - slant_range_ref
+        az_arr[index_arr] = aztime_displaced - aztime_ref
+
+    return rg_arr, az_arr
+
+
+def get_enu_vector_ecef(lon, lat, unit='degree'):
+    '''
+    Calculate the east, north, and up vectors in ECEF for lon / lat provided 
+
+    Parameters
+    ----------
+    lon: np.ndarray
+        Longitude of the points to calculate ENU vectors
+    lat: np.ndarray
+        Latitude of the points to calculate ENU vectors
+    unit: str
+        Unit of the `lon` and `lat`. Default is `degree`
+
+    Return
+    ------
+    vec_e: np.ndarray
+        unit vector of "east" direction in ECEF
+    vec_n: np.ndarray
+        unit vector of "north" direction in ECEF
+    vec_u: np.ndarray
+        unit vector of "up" direction in ECEF
+
+
+    '''
+    if unit=='degree':
+        lon_rad = np.deg2rad(lon)
+        lat_rad = np.deg2rad(lat)
+    elif unit=='radian':
+        lon_rad = lon
+        lat_rad = lat
+    else:
+        raise ValueError(f'"{unit}" was provided for "unit", '
+                         'which needs to be either "degree" or "radian"')
+
+    vec_u = np.array([np.cos(lon_rad) * np.cos(lat_rad),
+                       np.sin(lon_rad) * np.cos(lat_rad),
+                       np.sin(lat_rad)])
+
+    vec_n = np.array([-np.cos(lon_rad) * np.sin(lat_rad),
+                      -np.sin(lon_rad) * np.sin(lat_rad),
+                      np.cos(lat_rad)])
+
+    vec_e = np.cross(vec_n, vec_u, axis=0)
+
+    return vec_e, vec_n, vec_u
